refactor: Changed UsersGames to UserGame
