+++ src/main/java/org/harmony/endofline/userGame/UserGameRepository.java
@@ -1,0 +1,6 @@
+package org.harmony.endofline.userGame;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface UserGameRepository extends CrudRepository<UserGame, Integer> {
+}
