--- src/main/java/org/harmony/endofline/usersGames/UsersGamesRepository.java
@@ -1,6 +1,0 @@
-package org.harmony.endofline.usersGames;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface UsersGamesRepository extends CrudRepository<UsersGames, Integer> {
-}
